refactor: extract reusable BoardsComponentWrite and wire it into /boards/new and /boards/[boardId]/edit
