+++ 03-maincamp-triptalk/03-CNA/src/app/boards/[boardId]/edit/page.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+import BoardsComponentWrite from "@/components/boards-write"
+
+export default function BoardsEditPage() {
+  return <BoardsComponentWrite />
+}
